Move state manipulation code to thingtalk utils
Except for a few helpers that I need to get rid of. As part of
this change, state manipulation code is now updated to return
target states instead of full states, because that makes more
sense.
