+++ lib/transaction-dialogues/templates/refinement.genie
@@ -31,7 +31,6 @@ import { Ast, } from 'thingtalk';
 import * as C from '../../templates/ast_manip';
 import ThingpediaLoader from '../../templates/load-thingpedia';
 import * as D from '../dialogue_acts';
-import * as S from '../state_manip';
 
 search_result_preamble : D.ContextInfo = {
     ["{there are|i can see|i have found|i can recommend} {many|several} ${base[plural=other]} {matching your request|matching your constraints|with those characteristics|like that|}."]: (ctx:ctx_search_command, base:base_noun_phrase<functionName = ctx.currentFunction>)
@@ -41,7 +40,7 @@ search_result_preamble : D.ContextInfo = {
         => D.checkSearchResultPreamble(ctx, base, null, false);
 }
 
-system_generic_proposal : S.AgentReplyRecord = {
+system_generic_proposal : D.AgentReplyRecord = {
     ["{are you looking for a|how about a|how about the} ${proposal[plural=one]}"]: (ctx:ctx_search_command, proposal:answer_noun_phrase<functionName = ctx.currentFunction>)
         => D.makeRefinementProposal(ctx, proposal),
     ["{are you looking for|how about|how about} ${proposal}"]: (ctx:ctx_search_command, proposal:anything_phrase<functionName = ctx.currentFunction>)
